* Fixed a bug when it was impossible to input a cheat that had some of its keys assigned to a command * Major experimental change - bsp tree extended to 32 bits * Loading of extended node info in ZDoom format is supported (extra debugging statements are in for now) * Seg angle and offset are computed on the fly in case seg structure has zero value for them

+++ R_defs.h
@@ -268,7 +268,7 @@ typedef struct
 {
   fixed_t  x,  y, dx, dy;        // Partition line.
   fixed_t bbox[2][4];            // Bounding box for each child.
-  unsigned short children[2];    // If NF_SUBSECTOR its a subsector.
+  unsigned int children[2];      // If NF_SUBSECTOR its a subsector.
 } node_t;
 
 // posts are runs of non masked source pixels
